Extract SimpleGitManager from GitProcessor
To support future PooledGitManager
