+++ cf-ops-automation-broker-core/pom.xml
@@ -69,6 +69,13 @@
 		  <version>4.9.0.201710071750-r</version>
 		</dependency>
 
+        <!-- https://mvnrepository.com/artifact/org.apache.commons/commons-pool2 -->
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-pool2</artifactId>
+			<version>2.6.0</version>
+		</dependency>
+
 		<!-- Use SimpleLogger as the slf4j implementation in tests for getting debug traces from libs
 		<dependency>
 			<groupId>org.slf4j</groupId>
